requirements.txt
fastapi
uvicorn
python-dotenv
openai
motor
pymongo
passlib[bcrypt]
python-jose[cryptography]
python-multipart
